requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
discord.py==2.6.0
feedparser==6.0.11
frozenlist==1.7.0
idna==3.10
multidict==6.6.4
propcache==0.3.2
sgmllib3k==1.0.0
typing_extensions==4.14.1
yarl==1.20.1
yt-dlp==2025.8.11
